+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 -   Autocrafting Upgrade
+-   Display 'b' after fluid amount inside grid
 
 ### Changed